+++ util/ContinuousArray.h
@@ -37,7 +37,7 @@ namespace Core {
     private:
 
         ContinuousArray<T>::Interpolator getInterpolator();
-        void getInterpolationValuesForTValue(Real t, Int32& lowerIndex, Int32& upperIndex, Real& localT) {
+        void getInterpolationValuesForTValue(Real t, UInt32& lowerIndex, UInt32& upperIndex, Real& localT) {
             UInt32 tValueCount = this->tValues.size();
             if (tValueCount == 0) {
                 lowerIndex = upperIndex = -1;
@@ -53,8 +53,8 @@ namespace Core {
                 lowerIndex++;
                 upperIndex++;
             }
-            lowerIndex = Math::clamp(lowerIndex, 0, tValueCount - 1);
-            upperIndex = Math::clamp(upperIndex, 0, tValueCount - 1);
+            lowerIndex = Math::clamp(lowerIndex, (UInt32)0, tValueCount - 1);
+            upperIndex = Math::clamp(upperIndex, (UInt32)0, tValueCount - 1);
             Real lowerTValue = this->tValues[lowerIndex];
             Real upperTValue = this->tValues[upperIndex];
             localT = (t - lowerTValue) / (upperTValue - lowerTValue);
